% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds.R
\name{birds}
\alias{birds}
\title{Limit the dataset by given species and variables}
\usage{
birds(dat, UTFRREC, variables = NULL)
}
\arguments{
\item{dat}{A list of species given in scientific name format, common name, or 4 letter
bander abbreviation.}

\item{UTFRREC}{The dataset.}

\item{variables}{Two variables from dataset.}
}
\value{
A dataframe with the new dataset.
}
\description{
This loads the UTFRREC dataset and then selects
the rows based on given species, and can also limit columns based on given variables.
}
\examples{
dat <- c("CACH", "TUTI", "BLJA", "AMCR")
variables <- c("MSF", "Weight_oz")
birds(dat, UTFRREC, variables)
birds(c("CACH", "TUTI", "BLJA", "AMCR"), UTFRREC)
#export
}
