% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{paint}
\alias{paint}
\title{Visualize the variables given with painted trees}
\usage{
paint(dat, variables, UTFRREC)
}
\arguments{
\item{dat}{A list of species given in scientific name format, common name, or 4 letter
bander abbreviation.}

\item{variables}{Two variables from dataset.}

\item{UTFRREC}{The dataset.}
}
\value{
Two painted phylogenetic trees.
}
\description{
This selects data based on the given species and variables. The function will result in two
painted phylogenetic trees that correspond with the variables given. Discrete variables
require placing the legend.
}
\examples{
dat <- c("CACH", "TUTI", "BLJA", "AMCR")
variables <- c("MSF", "Weight_oz")
paint(dat, variables, UTFRREC)
paint(c("CACH", "TUTI", "BLJA", "AMCR"), c("MSF", "Weight_oz"), UTFRREC)
#export
}
