% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylocking.R
\name{phylocking}
\alias{phylocking}
\title{Select data, create trees, and find relationships between variables}
\usage{
phylocking(dat, variables, UTFRREC)
}
\arguments{
\item{dat}{A list of species given in scientific name format, common name, or 4 letter
bander abbreviation.}

\item{variables}{Two variables from dataset.}

\item{UTFRREC}{The dataset.}
}
\value{
Two painted phylogenetic trees.
}
\description{
This function selects data based on the given species and variables. The function will
create a plain phylogenetic tree, two painted trees, and will provide corrlelation or model
output for the given variables.
}
\examples{
dat <- c("CACH", "TUTI", "BLJA", "AMCR")
variables <- c("MSF", "Weight_oz")
phylocking(dat, variables, UTFRREC)
phylocking(c("CACH", "TUTI", "BLJA", "AMCR"), c("MSF", "Weight_oz"), UTFRREC)
#export
}
