% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.R
\name{relationship}
\alias{relationship}
\title{Find the relationship between given variables}
\usage{
relationship(dat, variables, UTFRREC)
}
\arguments{
\item{dat}{A list of species given in scientific name format, common name, or 4 letter
bander abbreviation.}

\item{variables}{Two variables from dataset.}

\item{UTFRREC}{The dataset.}
}
\value{
Two painted phylogenetic trees.
}
\description{
This selects data based on the given species and variables. The function chooses a model
to run depending on the chosen variables. If both variables are discrete, a correlation
matrix will be created with corHMM. If both variables are continuous, a correlation matrix
will be created with ape. If there is one continuous variable and one discrete variable,
a mixed model will be created with phylolm.
}
\examples{
dat <- c("CACH", "TUTI", "BLJA", "AMCR")
variables <- c("MSF", "Weight_oz")
relationship(dat, variables, UTFRREC)
relationship(c("CACH", "TUTI", "BLJA", "AMCR"), c("MSF", "Weight_oz"), UTFRREC)
#export
}
