% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{traits}
\alias{traits}
\title{Limit the dataset by variables}
\usage{
traits(variables, UTFRREC)
}
\arguments{
\item{variables}{Two variables from dataset.}

\item{UTFRREC}{The dataset.}
}
\value{
A dataframe with the new dataset.
}
\description{
This loads the UTFRREC dataset and then selects
the columns based on the given variables.
}
\examples{
variables <- c("MSF", "Weight_oz")
traits(variables, UTFRREC)
traits(c("MSF", "Weight_oz"), UTFRREC)
#export
}
