% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{Create a phylogenetic tree based on given species}
\usage{
tree(dat, UTFRREC)
}
\arguments{
\item{dat}{A list of species given in scientific name format, common name, or 4 letter
bander abbreviation.}

\item{UTFRREC}{The dataset.}
}
\value{
A plot with a phylogenetic tree.
}
\description{
This makes a tree for the given species with the Open Tree of Life simulated tree. There are
also estimated branch lengths for the tree.
}
\examples{
dat <- c("CACH", "TUTI", "BLJA", "AMCR")
tree(dat, UTFRREC)
tree(c("CACH", "TUTI", "BLJA", "AMCR"), UTFRREC)
#export
}
